Refactor util code and files
