+++ src/lib/persist.js
@@ -1,9 +1,11 @@
+import welcome from './welcome.txt';
+
 const PERSIST_KEY_PREFIX = 'js-playground-by-zill';
 
 const { NODE_ENV } = process.env;
 const storage = NODE_ENV === 'production' ? localStorage : sessionStorage;
 
-export function getPersist(key) {
+function getPersist(key) {
   try {
     const item = storage.getItem(`${PERSIST_KEY_PREFIX}-${key}`);
     return JSON.parse(item);
@@ -12,6 +14,25 @@ export function getPersist(key) {
   }
 }
 
-export function setPersist(key, value) {
+function setPersist(key, value) {
   storage.setItem(`${PERSIST_KEY_PREFIX}-${key}`, JSON.stringify(value));
+}
+
+const CONTENT_KEY = 'content';
+export function persistContent(text) {
+  setPersist(CONTENT_KEY, text.trim());
+}
+
+export function getPersistContent() {
+  const persistContent = getPersist(CONTENT_KEY) || '';
+  return `${persistContent.trim() || welcome.trim()}\n`;
+}
+
+const VIM_MODE_KEY = 'vim-mode-enabled';
+export function setPersistVimMode(mode) {
+  setPersist(VIM_MODE_KEY, mode);
+}
+
+export function getPersistVimMode() {
+  return getPersist(VIM_MODE_KEY) || false;
 }
